prepare for release with GHA (#27)
* prepare for release with GHA

* refactor examples for build
