+++ .gitignore
@@ -1,9 +1,11 @@
+.idea
+.code
 **/vendor/
 .pulumi
 **/bin/
 **/obj/
 Pulumi.*.yaml
-sdk/nodejs/node_modules
+**/node_modules/
 
 **/command-output/
 
@@ -12,3 +14,9 @@ sdk/nodejs/node_modules
 
 yarn.lock
 **/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
