--- .travis.yml
@@ -1,29 +1,0 @@
-# It may be tempting to add parens around each individual clause in this expression, but Travis then builds pushes anyway
-if: branch = master OR branch =~ ^features/ OR branch =~ ^feature- OR branch =~ ^release/ OR tag =~ ^v\d+.*
-language: go
-go: 1.14.x
-dist: xenial
-sudo: true # give us 7.5GB and >2 bursted cores.
-git:
-  depth: false
-services:
-  - docker
-before_install:
-    - docker pull consul
-    - docker run -d -p 8500:8500 --name=dev-consul consul agent -dev -client 0.0.0.0
-    - git clone https://github.com/pulumi/scripts ${GOPATH}/src/github.com/pulumi/scripts
-    - source ${GOPATH}/src/github.com/pulumi/scripts/ci/prepare-environment.sh
-    - source ${PULUMI_SCRIPTS}/ci/keep-failed-tests.sh
-install:
-    - source ${PULUMI_SCRIPTS}/ci/install-common-toolchain.sh
-    - source ${PULUMI_SCRIPTS}/ci/configure-aws.sh
-    - curl -L https://get.pulumi.com/ | bash
-    - export PATH=$HOME/.pulumi/bin:$PATH
-before_script:
-    - ${PULUMI_SCRIPTS}/ci/ensure-dependencies
-script:
-    - make travis_${TRAVIS_EVENT_TYPE}
-after_failure:
-    - ${PULUMI_SCRIPTS}/ci/upload-failed-tests
-notifications:
-    webhooks: https://zlmgkhmhjc.execute-api.us-west-2.amazonaws.com/stage/travis
