+++ examples/examples_nodejs_test.go
@@ -1,0 +1,34 @@
+// Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
+// +build dotnet all
+
+package examples
+
+import (
+	"path"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions()
+	baseJS := base.With(integration.ProgramTestOptions{
+		Config: map[string]string{
+			"consul:address": "http://127.0.0.1:8500",
+		},
+		Dependencies: []string{
+			"@pulumi/consul",
+		},
+	})
+
+	return baseJS
+}
+
+func TestAccNode(t *testing.T) {
+	test := getJSBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "node"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
