+++ examples/examples_test.go
@@ -4,20 +4,10 @@ package examples
 
 import (
 	"os"
-	"path"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
 )
-
-func TestAccNode(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "node"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
 
 func getCwd(t *testing.T) string {
 	cwd, err := os.Getwd()
@@ -32,18 +22,4 @@ func getBaseOptions() integration.ProgramTestOptions {
 	return integration.ProgramTestOptions{
 		ExpectRefreshChanges: true,
 	}
-}
-
-func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions()
-	baseJS := base.With(integration.ProgramTestOptions{
-		Config: map[string]string{
-			"consul:address": "http://127.0.0.1:8500",
-		},
-		Dependencies: []string{
-			"@pulumi/consul",
-		},
-	})
-
-	return baseJS
 }
