--- provider/cmd/pulumi-resource-consul/.gitignore
@@ -1,1 +1,0 @@
-schema.go
