--- sdk/dotnet/.gitignore
@@ -1,1 +1,0 @@
-version.txt
